+++ train_tssdnet.py
@@ -3,8 +3,8 @@ import argparse
 import pathlib
 
 import torch
-from utils.config.load_config import load_config_file
-from data_loaders.data_modules.SPCUP22DataModule import SPCUP22DataModule
+from SPCUP2022.utils.config import load_config_file
+from datasets.SPCUP22DataModule import SPCUP22DataModule
 from models.raw_audio.ResTSSDNet import ResTSSDNetWrapper
 from models.raw_audio.IncTSSDNet import IncTSSDNetWrapper
 import pytorch_lightning as pl
